requirements.txt
requests
python-dotenv
redis
openai
pytest
flask
flask-cors
datetime
python-dateutil
tweepy
discord
python-telegram-bot
gspread
oauth2client
schedule
tweepy
search_web
webdriver-manager
selenium
gunicorn
grpcio
grpcio-tools
protobuf
setuptools
PyNaCl